Combine a bunch of classes into XJoystick
