+++ tests/xbot/common/injection/UnitTestModule.java
@@ -9,7 +9,6 @@ import xbot.common.controls.actuators.XServo;
 import xbot.common.controls.actuators.XSolenoid;
 import xbot.common.controls.actuators.XSpeedController;
 import xbot.common.controls.sensors.XAnalogInput;
-import xbot.common.controls.sensors.XCompleteJoystick;
 import xbot.common.controls.sensors.XDigitalInput;
 import xbot.common.controls.sensors.XEncoder;
 import xbot.common.controls.sensors.XGyro;
@@ -85,8 +84,7 @@ public class UnitTestModule extends AbstractModule {
         
         this.install(new FactoryModuleBuilder()
                 .implement(XPowerDistributionPanel.class, MockPowerDistributionPanel.class)
-                .implement(XCompleteJoystick.class, MockJoystick.class)
-                //.implement(XCompleteJoystick.class, MockCompleteJoystick.class)
+                .implement(XJoystick.class, MockJoystick.class)
                 .implement(XEncoder.class, MockEncoder.class)
                 .implement(XDigitalInput.class, MockDigitalInput.class)
                 .implement(XAnalogInput.class, MockAnalogInput.class)
